1. 【新增】售后单售后原因类型支持搜索
2. 【新增】手工创建订单折扣可输入正数

3. 【优化】盘点申请单确认

4. 【修复】采购退货单模拟出库失败问题

5. 【新增】订单金额客户实付与结算金额

6. 【优化】仓库发货统计报表物料名称显示

7. 【优化】自有仓储虚拟发货逻辑

8. 【修复】基础物料分类管理问题

+++ app/ome/dbschema/order_extend.php
@@ -14,7 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 $db['order_extend'] = array(
     'columns' => array(
         'order_id'             => array(
@@ -152,7 +151,7 @@ $db['order_extend'] = array(
             'default' => '0',
         ),
         'es_time'              => array(
-            'type'    => 'tinyint',
+            'type'    => 'int',
             'label'   => '物运时间',
             'default' => '0',
         ),
